fixed incorrects coords, changed all int to s16 or u16, now wireframe and flat cube is drawing correct

+++ src/display.h
@@ -16,11 +16,11 @@ enum render_method {
 } render_method;
 
 void initialize_window(void);
-void draw_grid(int grid_size);
-void draw_pixel(int x, int y, u8 color);
-void draw_rect(int dx, int dy, int width, int height, u8 color);
-void draw_line(int x0, int y0, int x1, int y1, u16 color);
-void draw_triangle(int x0, int y0, int x1, int y1, int x2, int y2, u16 color);
+void draw_grid(s16 grid_size);
+void draw_pixel(s16 x, s16 y, u8 color);
+void draw_rect(s16 dx, s16 dy, s16 width, s16 height, u8 color);
+void draw_line(s16 x0, s16 y0, s16 x1, s16 y1, u16 color);
+void draw_triangle(s16 x0, s16 y0, s16 x1, s16 y1, s16 x2, s16 y2, u16 color);
 void draw_fix16(fix16 v);
 void draw_fix16_reset(void);
